Moved classes around to make the structure more understandable
