+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/core/OpModeExtended.java
@@ -4,12 +4,15 @@ import android.util.Log;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 
+import org.firstinspires.ftc.teamcode.core.iobuiltin.GamepadExtended;
+import org.firstinspires.ftc.teamcode.core.iobuiltin.LogRecorder;
+import org.firstinspires.ftc.teamcode.core.structure.ClassHolder;
+import org.firstinspires.ftc.teamcode.core.structure.Registry;
+import org.firstinspires.ftc.teamcode.core.tealisp.TealispFileManager;
 import org.firstinspires.ftc.teamcode.core.tealisp.TealispSubsystemInterface;
-import org.majora320.tealisp.evaluator.JavaInterface;
 import org.majora320.tealisp.evaluator.LispException;
 
 import java.io.File;
-import java.util.Set;
 
 import static org.apache.commons.lang3.exception.ExceptionUtils.getStackTrace;
