+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/core/iobuiltin/GamepadExtended.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.core;
+package org.firstinspires.ftc.teamcode.core.iobuiltin;
 
 import com.qualcomm.robotcore.hardware.Gamepad;
