+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/core/structure/ClassHolder.java
@@ -1,9 +1,9 @@
-package org.firstinspires.ftc.teamcode.core;
+package org.firstinspires.ftc.teamcode.core.structure;
 
+import org.firstinspires.ftc.teamcode.core.OpModeExtended;
 import org.majora320.tealisp.evaluator.JavaInterface;
 
 import java.util.Map;
-import java.util.Set;
 
 public abstract class ClassHolder {
     public OpModeExtended context;
