+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/core/structure/Setting.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.core;
+package org.firstinspires.ftc.teamcode.core.structure;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
