+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/core/tealisp/TealispFileManager.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.core;
+package org.firstinspires.ftc.teamcode.core.tealisp;
 
 import android.os.FileObserver;
 import android.support.annotation.Nullable;
