+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/core/tealisp/TealispSensorInterface.java
@@ -1,14 +1,10 @@
 package org.firstinspires.ftc.teamcode.core.tealisp;
 
-import android.util.Log;
-
-import org.firstinspires.ftc.teamcode.core.Registry;
+import org.firstinspires.ftc.teamcode.core.structure.Registry;
 import org.majora320.tealisp.evaluator.JavaInterface;
 import org.majora320.tealisp.evaluator.LispException;
 import org.majora320.tealisp.evaluator.LispObject;
 import org.majora320.tealisp.evaluator.StackFrame;
-
-import java.util.Arrays;
 
 public class TealispSensorInterface extends JavaInterface {
     @Override
