+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/testing/TestAuto.java
@@ -5,15 +5,14 @@ import android.util.Log;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
-import org.firstinspires.ftc.teamcode.core.ClassHolder;
+import org.firstinspires.ftc.teamcode.core.structure.ClassHolder;
 import org.firstinspires.ftc.teamcode.core.OpModeExtended;
-import org.firstinspires.ftc.teamcode.core.Registry;
-import org.firstinspires.ftc.teamcode.core.SensorManager;
-import org.firstinspires.ftc.teamcode.core.Subsystem;
+import org.firstinspires.ftc.teamcode.core.structure.Registry;
+import org.firstinspires.ftc.teamcode.core.structure.SensorManager;
+import org.firstinspires.ftc.teamcode.core.structure.Subsystem;
 
 import java.io.File;
 import java.util.HashMap;
-import java.util.Map;
 
 @Autonomous(name = "Test", group = "Tests")
 public class TestAuto extends OpModeExtended {
