+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/testing/TestHolder.java
@@ -1,19 +1,16 @@
 package org.firstinspires.ftc.teamcode.testing;
 
-import org.firstinspires.ftc.teamcode.core.ClassHolder;
+import org.firstinspires.ftc.teamcode.core.structure.ClassHolder;
 import org.firstinspires.ftc.teamcode.core.OpModeExtended;
-import org.firstinspires.ftc.teamcode.core.SensorInterface;
-import org.firstinspires.ftc.teamcode.core.SensorManager;
-import org.firstinspires.ftc.teamcode.core.Subsystem;
+import org.firstinspires.ftc.teamcode.core.structure.SensorManager;
+import org.firstinspires.ftc.teamcode.core.structure.Subsystem;
 
 import org.firstinspires.ftc.teamcode.core.tealisp.TealispSensorInterface;
 import org.firstinspires.ftc.teamcode.core.tealisp.TealispSubsystemInterface;
 import org.majora320.tealisp.evaluator.JavaInterface;
 
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
-import java.util.Set;
 
 public class TestHolder extends ClassHolder {
